+++ test/setup.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-alias show="valgrind -q --leak-check=yes $TESTDIR/../show"
+alias show="valgrind -q --leak-check=yes --show-leak-kinds=all $TESTDIR/../show"